+++ __init__.py
@@ -87,7 +87,10 @@ ui_modules = [
     "color_def", "sv_IO_panel", "sv_templates_menu", "sv_panels", "nodeview_space_menu",
     # bgl modules
     "viewer_draw", "viewer_draw_mk2", "nodeview_bgl_viewer_draw",
-    "index_viewer_draw"
+    "index_viewer_draw",
+    # show git info
+    "development"
+
 ]
 
 # modules and pkg path, nodes are done separately.
